Updating templates to support non-asit execution
The asit test driver template has dependencies on resources that
our role account cannot access. This should allow for the tests
to continue to execute with the asit template, as well as creates
a new template for Widevine's use that defines a widevine-tf-test
friendly set of options for resource dependencies.

Change-Id: Id8b6b8284fe45d10a5ff5ceacb36a9b052f7eb38