Disable DroidGuard to prevent provisioning conflicts
When TF runs our GTests, it runs them concurrently with other system
processes that are running on the device. One of these is DroidGuard,
which seems to check periodically for an invalid Provisioning state.
Several of our tests introduce this state intentionally, and then we
enter a race condition with DroidGuard for provisioning status, which
results in confusing test failure results.

Disabling DroidGuard while the tests are running prevents us from
fighting over provisioning status.

Change-Id: I6db01a277b42f44e99e3374a37c698bfcd754d0b