Run GetKeyDataNullPointer test on its own
Updating the oemcrypto test config to run the GetKeyDataNullPointer test
in its own test execution, as it tests a condition that could lead to a
segfault during the test, which causes all the other tests to be skipped.
Running on its own allows all the tests to run, as well as still letting
us catch the status that this test is failing.

Change-Id: I03ea2032e211dcb69ce292bb0d12f7ac4a79d023