Add AudioWaltTest
This is a generic test for audio related WALT tests.

The test would first connect the device to host and start
the permission granting script for WALT board.
Then the host would issue the corresponding audio test commands
, swtich the FTDI board making device connected to WALT,
wait for a configurable amount of time, switch the device
to connect back to host, pull the result csv files from WALT
and post-processing them.

And the config file for this test is also included.

For more information about WALT project:
https://github.com/google/walt

Change-Id: I6c49e6452d7f59331393dcdcafb7e326bb44c43d