add TestzipLaunchControlProvider
Bug:74200685
Test: sponge/920a7b38-b35c-41d7-91d9-e7317e2ec9af

Add a new LCP that utilizes artifacts from two independent builds.

Primary build: this is the build the test triggered from, test.zip
artifact will be downloaded and used. It'll also be used as the default
parameters when posting to dashboard (if not overwritten).

Device build: this is the build all artifacts (except test.zip) will be
downloaded from. Specifically, it is the build the device will be
flashed with.

Change-Id: Ia0bbc5c5183843c1419979f2a422014d7dff2a56